Clarify use-permissions-for-rows
Previous documentation for this config option said "Use permission nodes to determine the number of rows. By default, with this setting enabled, all players have six rows, unless otherwise denoted by permissions." This caused a massive loss of productivity in the support Discord and confusion among the meaning of the word "denote". Some people's vocabulary "in their world" may be different, and this commit accounts for that.